requirements.txt
fastapi
uvicorn
litellm
python-dotenv
pinecone-client
sentence-transformers
torch